Add license
I checked with Udacity about adding a license, since they did not have
one in the starter code repo that I forked to begin this project. I
received confirmation from them on 2018-07-10 that I could add a license
to this repo (which is almost entirely my own work). They just asked
that I continue to credit them in the README, which I'm happy to do.